+++ src/code/scene.hpp
@@ -8,7 +8,7 @@ class Scene {
         Scene(shared_ptr<Game>);
         virtual void loop(shared_ptr<Game>);
         virtual void render(shared_ptr<Game>);
-        virtual void mouseUp(shared_ptr<Game>, sf::Vector2i, int);
+        virtual void mouseUp(shared_ptr<Game>, sf::Vector2f, int);
         virtual void keyDown(shared_ptr<Game>, int);
         virtual void keyUp(shared_ptr<Game>, int);
 };
@@ -18,7 +18,7 @@ class SceneTitle : public Scene {
         SceneTitle(shared_ptr<Game>);
         void loop(shared_ptr<Game>) override;
         void render(shared_ptr<Game>) override;
-        void mouseUp(shared_ptr<Game>, sf::Vector2i, int) override;
+        void mouseUp(shared_ptr<Game>, sf::Vector2f, int) override;
         void keyDown(shared_ptr<Game>, int) override;
         void keyUp(shared_ptr<Game>, int) override;
 };
@@ -28,7 +28,7 @@ class SceneField : public Scene {
         SceneField(shared_ptr<Game>);
         void loop(shared_ptr<Game>) override;
         void render(shared_ptr<Game>) override;
-        void mouseUp(shared_ptr<Game>, sf::Vector2i, int) override;
+        void mouseUp(shared_ptr<Game>, sf::Vector2f, int) override;
         void keyDown(shared_ptr<Game>, int) override;
         void keyUp(shared_ptr<Game>, int) override;
 
@@ -39,7 +39,7 @@ class SceneBattle : public Scene {
         SceneBattle(shared_ptr<Game>);
         void loop(shared_ptr<Game>) override;
         void render(shared_ptr<Game>) override;
-        void mouseUp(shared_ptr<Game>, sf::Vector2i, int) override;
+        void mouseUp(shared_ptr<Game>, sf::Vector2f, int) override;
         void keyDown(shared_ptr<Game>, int) override;
         void keyUp(shared_ptr<Game>, int) override;
 };
