+++ src/code/scenefield.cpp
@@ -13,7 +13,7 @@ void SceneField::render(shared_ptr<Game> game) {
 
 }
 
-void SceneField::mouseUp(shared_ptr<Game> game, sf::Vector2i pos, int button) {
+void SceneField::mouseUp(shared_ptr<Game> game, sf::Vector2f pos, int button) {
 
 }
